feat(backend): get user data from Google

+++ backend/models/user.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type User struct {
 	gorm.Model
-	OAuthID    uint        `gorm:"unique" json:"oauth_id"`
+	OAuthID    int64       `gorm:"unique" json:"oauth_id"`
 	AvatarURL  string      `gorm:"size:255" json:"avatar_url"`
 	Name       string      `gorm:"size:255" json:"name"`
 	GitHubURL  string      `gorm:"size:255" json:"github_url"`
